fix bugs in the refresh shelter selection

+++ SafePoint.Data/Entities/Shelter.cs
@@ -14,6 +14,6 @@ namespace SafePoint.Data.Entities
 
         public decimal LocY { get; set; }
 
-        public decimal MaxCapacity { get; set; }
+        public int MaxCapacity { get; set; }
     }
 }
